Move checking of error to chat page
