+++ src/components/card/ChatWithUserCard.js
@@ -113,7 +113,9 @@ const ChatWithUserCard = ({
       let queryString = queries.join('&');
       const routeWithUpdatedChatId = `/chat/${chatId}${queryString.length > 0 ? `?${queryString}` : ''}`;
       setSelectedChatId(chatId);
-      router.push(`/chat/[chatId]`, routeWithUpdatedChatId, { shallow: true });
+
+      // use replace instead of push as we do not need to set the highlighting of chat as a new route in history
+      router.replace(`/chat/[chatId]`, routeWithUpdatedChatId, { shallow: true });
     }
   };
   return (
