+++ src/components/chat/pages/ChatPage.js
@@ -5,6 +5,7 @@ import ChatDialog from '../modules/ChatDialog';
 import api from '../../../../utils/api';
 import styled from 'styled-components';
 import { EMAIL_BAR_HEIGHT, NAVBAR_HEIGHT } from '../../../../utils/constants/navbar';
+import { useRouter } from 'next/router';
 import useMediaQuery from '@kiwicom/orbit-components/lib/hooks/useMediaQuery';
 
 const NoChatsContainer = styled.div`
@@ -18,6 +19,46 @@ const ChatPage = ({ user, chatId, postId, postType, isViewingChatsForMyPost }) =
   const [selectedChatId, setSelectedChatId] = useState(hasSelectedAChat ? chatId : null);
   const [isNewChat, setIsNewChat] = useState(chatId == null || typeof chatId == 'undefined');
   const [hasNoChatForOwnPost, setHasNoChatForOwnPost] = useState(true);
+  const router = useRouter();
+
+  // checks before loading page
+  useEffect(() => {
+    /**
+     * Check if already created a chat for a post that is not yours
+     */
+    if (!isViewingChatsForMyPost) {
+      api.chats
+        .getChatsForPost(postId)
+        .then((rawChat) => {
+          if (rawChat.docs.length > 0) {
+            const chat = rawChat.docs[0].data(); // assumption: should only have one chat since the chat is for another user's post
+            router.push(`/chat/[chatId]`, `/chat/${chat.chatId}?postId=${postId}&postType=${postType}`, {
+              shallow: true,
+            });
+          }
+        })
+        .catch(() => console.log('chat does not exist'));
+    }
+
+    if (chatId) {
+      api.chats
+        .getChat(chatId)
+        .then((rawChat) => {
+          // Check if chat exists
+          if (!rawChat.exists) {
+            router.push('/');
+          }
+          const chat = rawChat.data();
+          // Check if logged in user is part of the chat
+          if (chat.npo.id !== user.user.userId && chat.donor.id !== user.user.userId) {
+            router.push('/');
+          }
+        })
+        .catch((err) => {
+          console.log(err);
+        });
+    }
+  }, []);
 
   useEffect(() => {
     // only get chats for post if postId query is given
